refactor(fsd): i18n を shared/i18n へ移設 (#356)
A2 Lane2 shared foundation の i18n segment(Tier1・機械的移設)。
locales/contexts(I18n)/hooks(useTranslation) は FSD レイヤ外なので
shared/i18n へ集約し、FSD 命名(kebab)に合わせてリネーム。

- src/locales/* → src/shared/i18n/messages/
- src/contexts/I18nContext.tsx(+test) → src/shared/i18n/i18n-context.tsx
- src/hooks/useTranslation.ts → src/shared/i18n/use-translation.ts
- @/locales・@/contexts/I18nContext・@/hooks/useTranslation の import を
  @/shared/i18n/* へ書換(残存0)。テスト内の相対 ./I18nContext も追随。
- ThemeContext は shared/ui/theme(ui PR)へ、useFiscalYearDefault は
  entity/feature へ後回しのため contexts/hooks に一時残置。

ロジック diff ゼロ(27 files・28+/28-・全て import 行)・type-check 緑・
62 tests 緑・knip クリーン。挙動不変。
