+++ frontend/src/pages/bank-transactions/BankTransactionsPage.tsx
@@ -5,7 +5,7 @@ import type { BankTransaction } from '@/types'
 import { Icon, StatusBadge, Button, Card, DataTable, TableStateRow, Pager, FilterBar, FilterField, PageHead, SortableTh, nextSort } from '@/components/ui'
 import type { SortState } from '@/components/ui'
 import type { StatusMeta } from '@/components/ui'
-import { useTranslation } from '@/hooks/useTranslation'
+import { useTranslation } from '@/shared/i18n/use-translation'
 import { useFiscalYearDefault } from '@/hooks/useFiscalYearDefault'
 import { useRowCursor } from '@/components/keyboard'
 import { yen } from '@/shared/lib/format'
