+++ frontend/src/pages/login/LoginPage.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { login } from '@/api/endpoints'
 import { storeToken } from '@/api/client'
 import { Icon, Button, Notice, LogoMark } from '@/components/ui'
-import { useTranslation } from '@/hooks/useTranslation'
+import { useTranslation } from '@/shared/i18n/use-translation'
 
 export default function LoginPage() {
   const { t } = useTranslation()
