+++ frontend/src/shared/i18n/i18n-context.tsx
@@ -5,7 +5,7 @@ import {
   translate,
   type Locale,
   type MessageKey,
-} from '@/locales'
+} from '@/shared/i18n/messages'
 
 const STORAGE_KEY = 'locale'
